requirements.txt
aiofiles==23.2.1
aiogram==2.25.1
aiohttp==3.8.6
aiosignal==1.3.1
annotated-types==0.7.0
APScheduler==3.10.4
asgiref==3.8.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.9.1
certifi==2024.7.4
charset-normalizer==3.3.2
dill==0.3.8
et-xmlfile==1.1.0
frozenlist==1.4.1
idna==3.7
magic-filter==1.0.12
multidict==6.0.5
mysql-connector-python==9.0.0
openpyxl==3.1.5
pydantic==2.8.2
pydantic_core==2.20.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
typing_extensions==4.12.2
tzlocal==5.2
yarl==1.9.4
